Add remaining source code changes and test scripts
## Kernel Module Updates
- Update takakryptfs crypto.c with missing function declarations
- Fix takakryptfs.h header file with proper includes
- Resolve final compilation warnings and build issues

## Agent Implementation Updates
- Update agent.go with enhanced error handling
- Improve encryption_handler.go with better key management
- Fix protocol.go constants and message handling

## Testing Infrastructure
- Add manual-test.sh for manual system testing
- Add test-encryption-only.sh for crypto-only tests
- Add test-full-system.sh for complete integration tests
- Add test-takakrypt-full.sh for comprehensive testing
- Add run-test-commands.sh for automated test execution

## System Status
- All compilation issues resolved
- Test infrastructure complete
- Ready for full system integration testing

+++ pkg/agent/agent.go
@@ -464,7 +464,7 @@ func (w *Worker) processRequests() error {
 	// Send response back to kernel
 	if response != nil {
 		responseMsg := &netlink.Message{
-			Type:      msg.Header.Operation,
+			Type:      uint32(msg.Header.Operation),
 			Sequence:  msg.Header.Sequence,
 			Data:      response,
 			Timestamp: time.Now(),
